requirements.txt
appnope==0.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.6
django-extensions==1.9.1
djangorestframework==3.7.3
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jason==0.1.7
jedi==0.11.0
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
